+++ src/ReadonlyRecord.ts
@@ -139,3 +139,26 @@ export const pick = (
   ks: ReadonlyArray<string>,
 ): (<A>(y: RR.ReadonlyRecord<string, A>) => RR.ReadonlyRecord<string, A>) =>
   RR.filterWithIndex(elemV(Str.Eq)(ks))
+
+/**
+ * Omit a set of keys from a `ReadonlyRecord`. The value-level equivalent of
+ * the `Omit` type. For omitting from records with typed keys, instead look at
+ * the `ReadonlyStruct` module.
+ *
+ * @example
+ * import { omit } from 'fp-ts-std/ReadonlyRecord'
+ * import { pipe } from 'fp-ts/function'
+ *
+ * const remaining = pipe(
+ *   { a: 1, b: 'two', c: [true] },
+ *   omit(['b'])
+ * )
+ *
+ * assert.deepStrictEqual(remaining, { a: 1, c: [true] })
+ *
+ * @since 0.16.0
+ */
+export const omit = (
+  ks: Array<string>,
+): (<A>(y: Record<string, A>) => Record<string, A>) =>
+  RR.filterWithIndex(not(elemV(Str.Eq)(ks)))